requirements.txt
discord.py
pytz
quart
quart-discord
aiohttp
python-dotenv